backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Ivan Shebanov", email = "i.shebanov@g.nsu.ru"},
]
dependencies = ["fastapi>=0.126.0", "sqlalchemy>=2.0.45", "alembic>=1.17.2", "dotenv>=0.9.9", "psycopg[binary]>=3.3.2", "pyjwt>=2.10.1", "pwdlib[argon2]>=0.3.0", "uvicorn>=0.40.0", "python-multipart>=0.0.21", "minio>=7.2.20"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
start.cmd = "uvicorn src.main:server --reload"
start.env = {PYTHONPATH = "./src"}
makemig = "alembic revision --autogenerate -m \"{initial migration}\""
migrate = "alembic upgrade head"

exercises.cmd = "python ./src/backend/database/scripts/set_exercises.py"
exercises.env = {PYTHONPATH = "./src"}
